re-organizes course content page by combining components
